VAN-402 Remove hardcoded region and zone from cldcvr.yaml
This project will now use the provided var.GOOGLE_REGION TF variable to
select the region. Decided to make the GKE cluster multi-zone so that
zone will not be selected.

The default_max_pods_per_node was reduced from default of 110 to 20
because when the cluster was switched from single zone to regional, it
is using more network resources and exhausting the IP range.